+++ keyboards/bastardkb/charybdis/3x5/keymaps/nathancooke7/config.h
@@ -1,13 +1,5 @@
 #pragma once
 
-#define IGNORE_MOD_TAP_INTERRUPT
-#define TAPPING_TERM_PER_KEY
-#define TAPPING_TERM 220
-
-#define COMBO_TERM 30
-
-#define MACRO_TIMER 20
-
 #define CHARYBDIS_MINIMUM_DEFAULT_DPI 1200
 #define CHARYBDIS_MINIMUM_SNIPING_DPI 400
 #define CHARYBDIS_DRAGSCROLL_REVERSE_Y
@@ -19,9 +11,3 @@
 #define ENCODERS_PAD_B { B2 }
 #define ENCODER_RESOLUTIONS { 4 }
 #endif
-
-/* conserve space on MCU */
-#undef LOCKING_SUPPORT_ENABLE
-#undef LOCKING_RESYNC_ENABLE
-#define NO_ACTION_ONESHOT
-#define NO_MUSIC_MODE
