+++ users/nathancooke7/features/pointing.h
@@ -1,6 +1,8 @@
 // Copyright 2021 Christopher Courtney, aka Drashna Jael're  (@drashna) <drashna@live.com>
 // SPDX-License-Identifier: GPL-2.0-or-later
 
+#include "nathancooke7.h"
+
 report_mouse_t pointing_device_task_keymap(report_mouse_t mouse_report);
 void           matrix_scan_pointing(void);
 bool           process_record_pointing(uint16_t keycode, keyrecord_t* record);
