+++ users/nathancooke7/process_records.c
@@ -1,0 +1,69 @@
+// Copyright 2020 Christopher Courtney, aka Drashna Jael're  (@drashna) <drashna@live.com>
+// SPDX-License-Identifier: GPL-2.0-or-later
+
+#include "nathancooke7.h"
+#include "version.h"
+#include "caps_word.h"
+
+// Defines actions tor my global custom keycodes. Defined in drashna.h file
+// Then runs the _keymap's record handier if not processed here
+
+/**
+ * @brief Keycode handler for keymaps
+ *
+ * This handles the keycodes at the keymap level, useful for keyboard specific customization
+ */
+__attribute__((weak)) bool process_record_keymap(uint16_t keycode, keyrecord_t *record) { return true; }
+__attribute__((weak)) bool process_record_secrets(uint16_t keycode, keyrecord_t *record) { return true; }
+
+/**
+ * @brief Main user keycode handler
+ *
+ * This handles all of the keycodes for the user, including calling feature handlers.
+ *
+ * @param keycode Keycode from matrix
+ * @param record keyrecord_t data structure
+ * @return true Continue processing keycode and send to host
+ * @return false Stop process keycode and do not send to host
+ */
+
+bool process_record_user(uint16_t keycode, keyrecord_t *record) {
+  if (!process_caps_word(keycode, record)) { return false; }
+
+  static bool tAuNtTeXt = false;
+
+  if (tAuNtTeXt) {
+    if (record->event.pressed) {
+      if (keycode != KC_SPC) {
+        tap_code(KC_CAPS);
+      }
+    }
+  }
+
+  switch (keycode) {
+    case SARCASM:
+      if (record->event.pressed) {
+        tAuNtTeXt = !tAuNtTeXt;
+        if (tAuNtTeXt) {
+          tap_code(KC_CAPS);
+        }
+        else if (host_keyboard_leds() & (1<<USB_LED_CAPS_LOCK)) {
+          tap_code(KC_CAPS);
+        }
+      }
+      return false;
+    case KC_SECRET_1 ... KC_SECRET_2: // Secrets!  Externally defined strings, not stored in repo
+      if (!record->event.pressed) {
+        send_string_with_delay(secrets[keycode - KC_SECRET_1], MACRO_TIMER);
+      }
+      return false;
+
+    default:
+      return true;
+  }
+}
+    return process_record_keymap(keycode, record) && process_record_secrets(keycode, record);
+}
+
+__attribute__((weak)) void post_process_record_keymap(uint16_t keycode, keyrecord_t *record) {}
+void                       post_process_record_user(uint16_t keycode, keyrecord_t *record) { post_process_record_keymap(keycode, record); }
